+++ 2014_RTSS_federated_original/src/outputs_result.rs
@@ -29,8 +29,8 @@ mod tests {
         NodeData { id, params }
     }
 
-    fn create_high_utilization_dag() -> Graph<NodeData, f32> {
-        let mut dag = Graph::<NodeData, f32>::new();
+    fn create_high_utilization_dag() -> Graph<NodeData, i32> {
+        let mut dag = Graph::<NodeData, i32>::new();
         let n0 = {
             let mut params = HashMap::new();
             params.insert("execution_time".to_owned(), 4);
@@ -40,15 +40,15 @@ mod tests {
         let n1 = dag.add_node(create_node(1, "execution_time", 4));
         let n2 = dag.add_node(create_node(2, "execution_time", 3));
         let n3 = dag.add_node(create_node(3, "execution_time", 3));
-        dag.add_edge(n0, n1, 1.0);
-        dag.add_edge(n0, n2, 1.0);
-        dag.add_edge(n0, n3, 1.0);
+        dag.add_edge(n0, n1, 1);
+        dag.add_edge(n0, n2, 1);
+        dag.add_edge(n0, n3, 1);
 
         dag
     }
 
-    fn create_low_utilization_dag() -> Graph<NodeData, f32> {
-        let mut dag = Graph::<NodeData, f32>::new();
+    fn create_low_utilization_dag() -> Graph<NodeData, i32> {
+        let mut dag = Graph::<NodeData, i32>::new();
         let n0 = {
             let mut params = HashMap::new();
             params.insert("execution_time".to_owned(), 3);
@@ -58,13 +58,13 @@ mod tests {
         let n1 = dag.add_node(create_node(0, "execution_time", 3));
         let n2 = dag.add_node(create_node(1, "execution_time", 4));
 
-        dag.add_edge(n0, n1, 1.0);
-        dag.add_edge(n0, n2, 1.0);
+        dag.add_edge(n0, n1, 1);
+        dag.add_edge(n0, n2, 1);
         dag
     }
 
-    fn create_period_exceeding_dag() -> Graph<NodeData, f32> {
-        let mut dag = Graph::<NodeData, f32>::new();
+    fn create_period_exceeding_dag() -> Graph<NodeData, i32> {
+        let mut dag = Graph::<NodeData, i32>::new();
         let mut params = HashMap::new();
         params.insert("execution_time".to_owned(), 20);
         params.insert("period".to_owned(), 10);
